Minor code cleanup. Added GroupService. Setup common base service unit tests.
