+++ SecretSanta/src/SecretSanta.Domain/Services/GiftService.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using SecretSanta.Domain.Models;
@@ -6,14 +7,17 @@ namespace SecretSanta.Domain.Services
 {
     public class GiftService
     {
-        private ApplicationDbContext DbContext {get;}
+        private ApplicationDbContext DbContext { get; }
+
         public GiftService(ApplicationDbContext dbContext)
         {
-            DbContext = dbContext;
+            DbContext = dbContext ?? throw new ArgumentNullException(nameof(dbContext));
         }
 
         public Gift AddGiftToUser(int userId, Gift gift)
         {
+            if (gift == null) throw new ArgumentNullException(nameof(gift));
+
             gift.UserId = userId;
             DbContext.Gifts.Add(gift);
             DbContext.SaveChanges();
@@ -23,6 +27,8 @@ namespace SecretSanta.Domain.Services
 
         public Gift UpdateGiftForUser(int userId, Gift gift)
         {
+            if (gift == null) throw new ArgumentNullException(nameof(gift));
+
             gift.UserId = userId;
             DbContext.Gifts.Update(gift);
             DbContext.SaveChanges();
@@ -37,6 +43,8 @@ namespace SecretSanta.Domain.Services
 
         public void RemoveGift(Gift gift)
         {
+            if (gift == null) throw new ArgumentNullException(nameof(gift));
+
             DbContext.Gifts.Remove(gift);
             DbContext.SaveChanges();
         }
