+++ SecretSanta/test/SecretSanta.Domain.Tests/Services/GiftServiceTests.cs
@@ -1,7 +1,3 @@
-using Microsoft.Data.Sqlite;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SecretSanta.Domain.Models;
 using SecretSanta.Domain.Services;
@@ -9,48 +5,8 @@ using SecretSanta.Domain.Services;
 namespace SecretSanta.Domain.Tests.Services
 {
     [TestClass]
-    public class GiftServiceTests
+    public class GiftServiceTests : DatabaseServiceTests
     {
-        ILoggerFactory GetLoggerFactory()
-        {
-            IServiceCollection serviceCollection = new ServiceCollection();
-            serviceCollection.AddLogging(builder =>
-            {
-                builder.AddConsole()
-                    .AddFilter(DbLoggerCategory.Database.Command.Name,
-                               LogLevel.Information);
-            });
-            return serviceCollection.BuildServiceProvider().
-            GetService<ILoggerFactory>();
-        }
-
-        private SqliteConnection SqliteConnection { get; set; }
-        private DbContextOptions<ApplicationDbContext> Options { get; set; }
-
-        [TestInitialize]
-        public void OpenConnection()
-        {
-            SqliteConnection = new SqliteConnection("DataSource=:memory:");
-            SqliteConnection.Open();
-
-            Options = new DbContextOptionsBuilder<ApplicationDbContext>()
-                .UseSqlite(SqliteConnection)
-                .UseLoggerFactory(GetLoggerFactory())
-                .EnableSensitiveDataLogging()
-                .Options;
-
-            using (var context = new ApplicationDbContext(Options))
-            {
-                context.Database.EnsureCreated();
-            }
-        }
-
-        [TestCleanup]
-        public void CloseConnection()
-        {
-            SqliteConnection.Close();
-        }
-
         [TestMethod]
         public void AddGift()
         {
