+++ SecretSanta/test/SecretSanta.Domain.Tests/Services/GroupServiceTests.cs
@@ -1,6 +1,6 @@
 ﻿namespace SecretSanta.Domain.Tests.Services
 {
-    public class UserServiceTests : DatabaseServiceTests
+    public class GroupServiceTests : DatabaseServiceTests
     {
         
     }
