+++ SecretSanta/test/SecretSanta.Domain.Tests/Services/ServiceTests.cs
@@ -1,0 +1,52 @@
+﻿using Microsoft.Data.Sqlite;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using SecretSanta.Domain.Models;
+
+namespace SecretSanta.Domain.Tests.Services
+{
+    public abstract class DatabaseServiceTests
+    {
+        private static ILoggerFactory GetLoggerFactory()
+        {
+            IServiceCollection serviceCollection = new ServiceCollection();
+            serviceCollection.AddLogging(builder =>
+            {
+                builder.AddConsole()
+                    .AddFilter(DbLoggerCategory.Database.Command.Name,
+                        LogLevel.Information);
+            });
+            return serviceCollection.BuildServiceProvider().
+                GetService<ILoggerFactory>();
+        }
+
+        private SqliteConnection SqliteConnection { get; set; }
+        protected DbContextOptions<ApplicationDbContext> Options { get; private set; }
+
+        [TestInitialize]
+        public void OpenConnection()
+        {
+            SqliteConnection = new SqliteConnection("DataSource=:memory:");
+            SqliteConnection.Open();
+
+            Options = new DbContextOptionsBuilder<ApplicationDbContext>()
+                .UseSqlite(SqliteConnection)
+                .UseLoggerFactory(GetLoggerFactory())
+                .EnableSensitiveDataLogging()
+                .Options;
+
+            using (var context = new ApplicationDbContext(Options))
+            {
+                context.Database.EnsureCreated();
+            }
+        }
+
+        [TestCleanup]
+        public void CloseConnection()
+        {
+            SqliteConnection.Close();
+        }
+    }
+}
